Support background color shorthand specified by CSS color string
